move a pagination test to test_image module
